Move and degrok @@start view class
